Remove the bits for the discrete logarithms
They were intended to stop a prover from including a point multiple times,
which had the soundness of questioned. Because the prover can have the point
whose discrete logarithm is being proved for be a power of two, we cannot
prevent a point from being included multiple times without asserting inequality
of the resulting point from all powers of two. This would cause a complete loss
in the performance gained from the divisors and cause us to revert to
incomplete addition.

On soundness, https://eprint.iacr.org/2022/596.pdf 3.2.1 explicitly claims
_distinct_ points sharing a z coordinate is invalid, not points sharing a z
coordinate in general. Accordingly, this should be within the claims of the
paper.

We do need the representation of the discrete logarithm to be consistent. By
making a supposed bit 2, it forces inclusion of that power of 2 multiple times.
If the representation is reused with a distinct generator, it will force the
distinct generator's power of two included the same amount of multiple times.
Accordingly, it is a consistent representation when used against tables formed
with this exact process and no further assumptions on the format of the
representation are made.

This makes the discrete logarithm proofs just 16 in-circuit multiplications, or
80 in total for the first layer (o T, i U, i V, r_i T, c G). With Seraphis, the
linking tag format, we could eliminate 4 of these, to save 16 multiplications
(if we do not achieve aggregation of claims as we did for set membership).

Due to how cheap discrete log proofs now are, if this holds (one of the first
topics for the earmarked fund), we can discuss a 16-layer tree with 16 items
each. This would use 512 bytes (not 128) for the vector commitments, yet only
take 16 * 16 multiplications for set membership and 16 * 16 multiplications for
discrete logarithms (+ (80-16) for the first layer, + whatever overhead). This
means we'd be discussing ~600 gates for entire proof (20% of the current proof,
near equivalent to an 8-output Bulletproof).

That performance sounds unbelievable so I don't want to argue it holds. It's
obvious more work on the soundness of elliptic curve divisors must be done. Not
only will that be a priority for the earmarked fund, we do have a fallback
that's twice as slow as divisors were prior expected to be. That's tolerable.