Modernize a bit
Removes references to potentially migrating to Seraphis. Removes reference to
JAMTIS (closing #2) as we're moving forward with CARROT and the possibility of
JAMTIS in the future is sufficiently-known. This text was expository to
convince people of moving forward with FCMP++.

Specify the range proofs are no longer hashed for the signature.

Specify only a single GBP is used for all inputs.