Hash the y coordinates on the leaf layer
The FCMP didn't prove OIC \in OutputSet.
It proved (-)O (-)I (-)C \in OutputSet. We'd have to prove the entire
composition so acceptable, when we've already contracted review on the
composition/SAL proof.

While we have a solution for permissibility re: branches, it doesn't apply
unless an opening over distinct generators is proven. That doesn't occur within
the FCMP. It does occur within the composition as a whole for OC, but the FCMP
is expected to work as the unit it is. Then I is a private point and cannot be
so opened.

We either have to redefine the composition, cheat in the expectation the SAL
will always open OC, reintroduce permissibility, or make the choice here.
Permissibility itself is painful and still doesn't work for I. No matter what,
the leaf layer will grow by at least 33% without redefining the already
worked-on composition.

This choice grows it by 100% as it's the most direct/proper solution. The hope
is the planned FCMP contest, and potential further optimizations, correct for
it.

This does avoid input malleation and the redefinition/migration of key images.